% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation function.R
\name{year_corr}
\alias{year_corr}
\title{Correlation function}
\usage{
year_corr(year)
}
\description{
This function allows you to select the correlation matrix for all numeric variables in the data for that year.
}
\examples{
year_corr()
}
