% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Age function.R
\name{year_oldest}
\alias{year_oldest}
\title{Age function}
\usage{
year_oldest(year)
}
\description{
This function allows you to select the oldest player in the NBA data for that year.
}
\examples{
year_oldest()
}
\keyword{age}
