% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Point function.R
\name{year_point}
\alias{year_point}
\title{Point function}
\usage{
year_point(year)
}
\description{
This function allows you to select the player with highest point in the NBA data for that year.
}
\examples{
year_point()
}
\keyword{points}
